rewrite source-map-cache tests
Use source-map-fixtures to get a fixture file with an inline source map. Add a
script to generate a coverage report for this specific fixture. Rewrite tests to
use the fixture, without (too much) hardcoding of values.

This will make it easier to upgrade the fixture in the future, as well as
regenerate the coverage report.