adapt remap-istanbul's mapping logic for nyc
remap-istanbul's mapping function seems to cover more edge cases, as well as
preserving "skip" annotations. This commit rewrites nyc's mapping logic to use a
helper function adapted from remap-istanbul. See
<https://github.com/SitePen/remap-istanbul/blob/30b67ad3f24ba7e0da8b8888d5a7c3c8480d48b2/lib/remap.js#L55:L108>
for the original.

Updated source-map-fixtures to get a fixture that includes an `/* istanbul
ignore if */` comment. Regenerated the coverage fixture and changed the
invocation of the 'branching' fixture so it has full coverage.

Fixes #67.