rework options processing
The use of lodash `_.extend(this, opts)`, meant that we had to work hard to avoid name collisions. The best example is using `tmpDirectory` as a method name to avoid conflicts with the `tempDirectory` option.
This also reduces our dependency on lodash, which is an eventual goal since lodash is very expensive to require.