shortcut include glob if not provided
Previously, the `include` glob defaulted to `**` or "accept everything".

Now, instead of creating a glob, we just bypass it and check only the `exclude` glob.

It is a fairly minor optimization, but it is in our require hook which gets executed tens of thousands of times, even in small test suites.

~20% reduction in our "coverage penalty" for the AVA test suite.