apply source maps to original report
Source maps are applied to a report that was just read from disk. Cloning is
needed for the tests but otherwise unnecessary.

Change path-related tests to look for the path property on the report, without
comparing the report contents. That's silly now that applySourceMaps() doesn't
do the cloning.

No need to copy coverage.b values when mapping branches.