Refactor source-map-cache to support caching
This is @novemberborns work, so I need to defer to him if extra explanation is needed.

apply source maps to original report

Source maps are applied to a report that was just read from disk. Cloning is
needed for the tests but otherwise unnecessary.

Change path-related tests to look for the path property on the report, without
comparing the report contents. That's silly now that applySourceMaps() doesn't
do the cloning.

No need to copy coverage.b values when mapping branches.

reword 'coverage' in source-map-cache

Use 'report' and fileReport' to be more consistent with index.js.

remove add() method from source-map-cache

It was only used by the tests, which can use addMap.

remove opts from SourceMapCache()

This is no longer used. This was also the last runtime dependency on lodash,
which has now been moved to the dev dependencies.