add test with contention for the cache
It spawns `cache-collision-runner`, which in turn spawns a number of `cache-collision-worker` processes.
It then coordinates all processes to try accessing the same cache file at the same time.
The end result is that they all see a cache miss, so they all attempt to write the cache entry themselves.
Turns out this doesn't throw any errors, and everything "just works".