run the tests twice. with and without cache
Update the `build-tests` script to run each test two times in a row,
once with caching, once without.

There is still one test that apparently requires state information from a previous test in order to pass.
I have made an ugly hack to get it working for now, and opened issue #112 to address correctly in the near future.