Record the hash in `hashCache` whether there was a cache miss or not.
Previously, we were only recording the hash in `hashCache` if there was a cache miss.

This meant that the `contentHash` property was only properly attached for files not in the cache, and that source maps would not be applied correctly to reports generated from cached instrumentations.

The solution is to always store the hash, but selectively attach `contentHash` to the reports based on `options.enableCache`.