Use find-cache-dir to determine cache directory.
Uses `find-cache-dir` for consistent cache locating with AVA, etc. It takes the extra step of ensuring a `package.json` file exists (and hunting up the directory tree for it), otherwise it returns null. If it can't find a suitable cache directory, caching is disabled.

I also dropped the `cacheDirectory()` method in favor of just setting a property.