Upgrade foreground-child, remove unnecessary arg munging
Foreground-child is now using cross-spawn-async in order to be able to
properly spawn shebangs and .cmd/.exe files.  This avoids the extra
cmd.exe that win-spawn throws into the mix, since that causes problems
when there is no PATH environ and a binary is being called explicitly
via a full path name.

Cross-spawn-async is a lot more code than win-spawn, but the approach it
takes is more surgical and well-tested.

Because of this, it's no longer necessary to unshift the node/io.js
binary onto the argument list when spawning a shebanged javascript file.
So, doing 'nyc mocha ...' on Windows will now work, and it's no longer
necessary to do 'nyc ./node_modules/mocha/bin/mocha.js ...' or any other
extra manual path munging.