Use `nyc` instead of `config.nyc` in package.json
Fixes #131

This also uses `pkg-up` to locate the nearest `package.json` from `cwd`, and will move `cwd` to there. This allows you to invoke `nyc` from a subdirectory of your project and still have it behave as expected.  This is something our caching mechanism already does. If it can't find a `package.json`, it just uses the original `cwd` option.