Split wrapping out from main nyc bin
This is a cleaner solution to the issue caused by spawning the nyc bin
from a test that is currently covered.

Also avoids having competing arg-parsing between yargs (better for
program arguments, user-flags, and such) and spawn-wrap, which just
splices off a set count of positional arguments from the array.