only include 'skip' if true in the original report
If present, `skip` should only be `true`. Copy it to the mapped location only if
that's the case.

This avoids setting `skip` to `undefined` while mapping the report. Note that
JSON.stringify() is used when writing the report to disk, so the undefined
values wouldn't have ended up in the reports before.