Add an EditorConfig and clean up readme
Note: introducing these changes raises a question in the coveralls
example at the bottom: does `nyc` really need to be run again for the
report to be available? and wouldn’t it be better just to have the lcov
report generated alongside the regular text display, and have the
`after_success` command just `cat` and pipe it to `coveralls`?