Add produceSourceMap option
This allows the user to control the source map output post instrumentation.

Not fully straightforward right now, but when combined with a fork of source-map-support that has inline support (or placing the files in the proper location), end to end covered source maps can be achieved.

For runtime source-map-support, see https://github.com/evanw/node-source-map-support/pull/118. Will be forking to @kpdecker/source-map-support soon, since the canonical project appears to be dead.