Allow eager instantiation of instrumenter
Some mocking libraries hook require and warn when there is a required
module which has not been explicitly registered (e.g. mockery).  It is
standard procedure to enable these mock hooks in a unit test prior to
requiring the code under test.

With lazy instantiation of the instrumenter, all of the instrumenter
library's `require`s generate warnings at test time.

This change adds an `--eager` option to nyc which forces instantiation
of the instrumentation library prior to walking any of the files, thus
eliminating any warnings from mockery.