fix: accept npm v5 package version
npm v5 transforms the version field from the simple number to an URL of
the tgz file in the package.json when installing nyc as a dependency.
The version is appended to the hash for cache comparison which is also
used to create a cache directory on the filesystem. The URL may include
illegal character and the creation of the file fails.