Add possibility to filter coveragemaps
By filtering the coveragemaps for certain files, using include and exclude, it is possible to gain some more control over the files being included in the report. Normally this is done when instrumenting the code, but it may be beneficial to create different reports and instrument the code only once.