fix: Enable es-modules by default. (#889)
A previous fix for #796 added an es-modules option and set it false by
default.  This default was a change of behavior in NYC as previously all
sources were parsed as ES modules by the instrumenter.  I believe the
old behavior is the correct default because it allows NYC to work with
modern code.  Code which uses certain features which violate strict mode
will have to disable es-modules.  This will not effect ES5 which follows
the rules of 'use strict'.