Fix missing command temp-directory
After changing tempDirectory to tempDir, with a fallback for tempDirectory, the default values set in the 'temp-dir' yargs option were being used rather than falling back to tempDirectory and then the default temp dir.
This change removes the yargs default value for the option 'temp-dir'.
This means the statement that assigns '_tempDirectory' in 'index.js' attempts to use temp-dir, then temp-directory, then the default value.  
This is now under test with some new test cases that show:
  - tempDir is preferred to tempDirectory
  - tempDirectory is used if tempDir isn't set
  - that we fall back to the '.nyc_output' dir if neither tempDir or tempDirectory are set.