Support and tests for nyc instrument command option 'clean'
Adds support for a 'clean' function for the instrument command, this will remove the 'instrument' output directory by default.
This behaviour can be switched off by specifying '--clean false'.
This option will do nothing if no output directory is specified.
This is a breaking change, the current behaviour is that an existing output directory is appended to.