Implement '--clean' option for nyc instrument command
With this change you can now tell nyc to clean the output directory when instrumenting, to do this an output directory must be specified.
I've currently set the default value for this to false so it can be a non-breaking change, although the default behaviour should be changed to true in the next major release.
I've included tests for this behaviour, and introduced lint support for `afterEach` and required the `makeDir` module