[skip ci]docs: draft for project root directory and `--cwd` doc
So I got a little off track while trying to add `instrument --cwd`, so I figured I'd spend some time understanding what it does, opposed to what I think it should do, and put some documentation together in the process.  I'm still not 100% sure this is correct, it's my reading of the intention of the code.  A few notes, I haven't mentioned `env.NYC_ENV` as I figured that it's not really supposed to be directly accessed by users.  I also didn't mention that `cwd` is fed in while creating a new instrumenter, as it looks to be ignored.