`nyc instrument` copies all files to an output directory, then adds instrumented versions where needed
So I've implemented this using `fs-extra`, not sure if that's the module you wish to go with here.
This is now relying on `testExclude.globSync` to get the correct files for instrumentation, so no more `shouldInstrument`
There's quite a few test changes in here, as `fs-extra.copySync` won't let you copy from `./` to `./outputMkdirHere`, but they're probably for the best anyway.
This also means that any attempt to store the instrumentation result in a subdir of the input dir will fail.