nyc instrument now copies all files, with permissions, from `input` to `output` dir
It ignores files in the `.git` folder and copies empty directories
Unfortunately I had to keep `fs-extra`, although this should be easy to drop when the project moves to node 8+ when `fs.copyFileSync` is made available.
We need to use `fs.chmodSync` to set the mode as any `write` mode flags get filtered when trying to set them through `fs.writeFileSync`.

What's missing?
 * No check to see where the working directory is relative to the project root directory
 * Tests confirming includes/excludes can be set in a config file or stanza and are observed by the instrument command
 * Tests confirming file permissions match between files in the input and output directories
 * Tests confirming we can't instrument outside of the project root directory