chore: Remove test artefacts that ensure failure on subsequent tests
If the tests fail with a bad 'test/fixtures/cli/foo-cache', then subsequent tests will also fail unless the cache is cleaned.
A failed test run also leaves a lot of files in 'test/temp-dir-nyc-integration' that are cleaned on success, but not before the start of the next run in case of failure.