fix: Avoid `source-map` module during `signal-exit` handler (#1191)
Prior to this change child processes would call upon the source-map
module to remap coverage during the `signal-exit` handler.  This was a
blocker for updating `source-map` to 0.7.x which uses an async
constructor method.

Now we simply save the source-map with the raw coverage data for each
file.  This allows the main nyc process to perform source-map remapping.
The benefit is that the main nyc process can use async code.