docs: link to next tracking issue in changelog
Information about the next version in development, which is published to the NPM registry without git version tags and without CHANGELOG entries, is tracked in an open Github issue. This is not very discoverable on its own. This makes it more discoverable but linking to it from the master CHANGELOG.

Closes #1236.